refactor(worker): extract media jobs + ffmpeg rendering into media_jobs.rs (sc-1636)
Gather the media cluster out of lib.rs into media_jobs.rs: the frame-extract,
person-detect, person-track and timeline-export job handlers (incl.
run_timeline_export_job) plus the ffmpeg rendering helpers (render/mux
segments, run_ffmpeg, crossfades, TimelineSegment/RenderSpec, the
person-detect candidate/track helpers, and export_request_from_job).

Both media regions move together so TimelineExportRequest and its
export_request_from_job builder stay co-located. Same pattern:
`use super::*;` (job-lifecycle helpers, path/payload utils, install
markers, ApiClient all resolve from the root/other modules unchanged),
pub(crate) on items reached from the dispatcher + tests, glob re-import
at the root, worker tests.rs imports repointed to super::media_jobs.
Behavior-preserving. lib.rs 3,346 -> 1,829 lines.
