+++ crates/sceneworks-worker/src/tests.rs
@@ -18,6 +18,9 @@ use super::gpu::{
     cpu_gpu, cpu_worker_id, fallback_gpu, gpu_worker_id, parse_nvidia_smi_gpus, visible_gpu_ids,
     worker_capabilities_with_utility,
 };
+use super::media_jobs::{
+    candidate_people, concat_file_contents, crossfade_duration, output_dimensions, run_ffmpeg,
+};
 use super::model_jobs::{
     check_downloaded_model_family, finalize_converted_dir, DownloadFamilyCheck,
 };
@@ -26,13 +29,12 @@ use super::supervisor::{
     utility_worker_specs, SupervisedChild, WorkerSpec,
 };
 use super::{
-    allow_pattern_matches, bounded_tail, candidate_people, cleanup_uploaded_import_source,
-    concat_file_contents, copy_lora_source, crossfade_duration, download_lora_source_url,
-    download_progress_payload, download_snapshot, fresh_asset_id, import_lora_source_path,
-    now_rfc3339, output_dimensions, resolve_model_convert_output, resolve_model_import_target,
-    run_ffmpeg, safe_download_dir, safe_project_path, value_f64, write_model_install_marker,
-    DownloadContext, DownloadProgress, HuggingFaceSnapshot, JsonObject, Settings, SnapshotFile,
-    WorkerError, DEFAULT_MAX_LORA_URL_BYTES, DEFAULT_MAX_MODEL_URL_BYTES,
+    allow_pattern_matches, bounded_tail, cleanup_uploaded_import_source, copy_lora_source,
+    download_lora_source_url, download_progress_payload, download_snapshot, fresh_asset_id,
+    import_lora_source_path, now_rfc3339, resolve_model_convert_output,
+    resolve_model_import_target, safe_download_dir, safe_project_path, value_f64,
+    write_model_install_marker, DownloadContext, DownloadProgress, HuggingFaceSnapshot, JsonObject,
+    Settings, SnapshotFile, WorkerError, DEFAULT_MAX_LORA_URL_BYTES, DEFAULT_MAX_MODEL_URL_BYTES,
     DEFAULT_TRANSITION_DURATION_SECONDS, INSTALL_MARKER,
 };
